+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Changelog
 
 ### Improvements and Changes
 
--   Added some type hints to the `Program` class (@rht, gh-1115).
+-   Type hints have been added to the `Program` class (@rht, gh-1115, gh-1134)
 -   Use [Black](https://black.readthedocs.io/en/stable/index.html) for code style
     and enforce it (along with a line length of 100) via the `style` (`flake8`)
     and `formatcheck` (`black --check`) CI jobs (@karalekas, gh-1132).